[#7] Make code more clean with extension and separate class
