+++ sample/src/main/java/co/nimblehq/sample/ui/login/LoginViewModel.kt
@@ -27,8 +27,8 @@ class LoginViewModel : ViewModel() {
             withContext(Dispatchers.IO) {
                 try {
                     val result = authenService.loginEmail(request)
-                    SurveyApi.instance.setTokenKey(result.accessToken ?: "")
-                    SurveyApi.instance.setTokenType(result.tokenType ?: "")
+                    SurveyApi.instance.setTokenKey(result.accessToken.orEmpty())
+                    SurveyApi.instance.setTokenType(result.tokenType.orEmpty())
                     _loginResult.postValue(Result.Success("Hello"))
                 } catch (exception: Exception) {
                     exception.printStackTrace()
