+++ survey-sdk/src/main/java/co/nimblehq/survey/sdk/constant/Constant.kt
@@ -1,0 +1,5 @@
+package co.nimblehq.survey.sdk.constant
+
+object Constant {
+    const val AUTHORIZATION_HEADER = "Authorization"
+}
